清空并重构数据库，实现增删改需要权限

+++ routes/api.js
@@ -4,11 +4,17 @@ var Note = require('../model/note').Note;
 
 /* GET users listing. */
 router.get('/notes', function(req, res, next) {
+  console.log('/notes...')
   Note.findAll({raw:true}).then((notes)=>{
+    console.log(notes)
     res.send({status: 0, data: notes})
   })
 });
 router.post('/notes/add',function(req, res, next){
+  if(!req.session.user){
+    return res.send({status: 1,errorMsg: '请先登录'})
+  }
+
   var note = req.body.note
   Note.create({text:note}).then(()=>{
     res.send({status:0})
@@ -18,6 +24,9 @@ router.post('/notes/add',function(req, res, next){
   console.log('add..', note)
 });
 router.post('/notes/edit',function(req, res, next){
+  if(!req.session.user){
+    return res.send({status: 1,errorMsg: '请先登录'})
+  }
   Note.update({text: req.body.note}, {where: {id: req.body.id}} ).then(()=>{
     console.log(arguments)
     res.send({status: 0})
@@ -27,6 +36,9 @@ router.post('/notes/edit',function(req, res, next){
   )
 });
 router.post('/notes/delete',function(req, res, next){
+  if(!req.session.user){
+    return res.send({status: 1,errorMsg: '请先登录'})
+  }
   Note.destroy({where:{id:req.body.id}}).then(()=>{
     res.send({status: 0})
     console.log('false')
